Try to set the Cookies for Comments image in the middle of the document.
The CFC image loads by default in the footer, howver for large-sized HTML
pages, it is possible for the image not to be loaded and attempts to login
will result in a 403.

By moving the CFC image higher up the HTML document, there is a better
chance at the image being loaded.